# openapi3.yaml
openapi: 3.0.1
info:
  title: Job Service
  description: Job Service
  version: 1.0.0
paths:
...
  /jobs/{jobId}/tasks:
    parameters:
      - $ref: '#/components/parameters/jobId'
    get:
      operationId: getTasks
      tags:
        - tasks
      summary: Get all the tasks of a job
      responses:
        200:
          description: Array of tasks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/tasksResponse'
        404:
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorMessage'
    post:
      operationId: createTask
      tags:
        - tasks
      summary: Create new task for job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/createTaskPayload'
                - type: array
                  items:
                    $ref: '#/components/schemas/createTaskPayload'
      responses:
        201:
          description: Task created successfully
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/createTaskResponse'
                  - $ref: '#/components/schemas/createMultipleTasksResponse'
        400:
          description: Invalid task input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorMessage'
        404:
          description: No such job in the database
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorMessage'
        409:
          description: An ingestion process with the same resourceId/version is already in progress (unique violation).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorMessage'
  /jobs/{jobId}/tasks/{taskId}:
    parameters:
      - $ref: '#/components/parameters/jobId'
      - $ref: '#/components/parameters/taskId'
    get:
      operationId: getTask
      summary: Get task by task ID
      responses:
        200:
          description: Requested task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/taskResponse'
        404:
          description: No such task in the database
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorMessage'
      tags:
        - tasks
    put:
      operationId: updateTask
      summary: Update task by task Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/updateTaskPayload'
      responses:
        200:
          description: Update task data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/defaultOkMessage'
              example:
                code: 'TASK_UPDATED_SUCCESSFULLY'
        400:
          description: Invalid request, could not update task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorMessage'
        404:
          description: Job/Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorMessage'
      tags:
        - tasks
    delete:
      operationId: deleteTask
      summary: Delete an existing task
      responses:
        200:
          description: task deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/defaultOkMessage'
              example:
                code: 'TASK_DELETED_SUCCESSFULLY'
        404:
          description: Job/Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorMessage'
      tags:
        - tasks
...
  /tasks/find:
    post:
      operationId: findTasks
      summary: return list of matching tasks
      tags:
        - tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/findTaskRequest'
      responses:
        200:
          description: Retrieved matching Tasks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/tasksResponse'
        400:
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorMessage'
        404:
          description: tasks not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorMessage'
...
  /tasks/findInactive:
    post:
      operationId: findInactiveTasks
      summary: retrive list of inactive task ids
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/findInactiveTasksRequest'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/taskIdList'
        400:
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorMessage'
      tags:
        - tasksManagement
...
components:
  parameters:
    jobId:
      in: path
      name: jobId
      description: ID of Job
      required: true
      schema:
        $ref: '#/components/schemas/jobId'
    taskId:
      in: path
      name: taskId
      description: ID of requested task
      required: true
      schema:
        $ref: '#/components/schemas/taskId'
...
  schemas:
    findTaskRequest:
      type: object
      description: task find model
      properties:
        id:
          $ref: '#/components/schemas/taskId'
        percentage:
          $ref: '#/components/schemas/percentage'
        jobId:
          $ref: '#/components/schemas/jobId'
        creationTime:
          $ref: '#/components/schemas/creationTime'
        updateTime:
          $ref: '#/components/schemas/updateTime'
        description:
          $ref: '#/components/schemas/description'
        type:
          $ref: '#/components/schemas/type'
        reason:
          $ref: '#/components/schemas/reason'
        status:
          $ref: '#/components/schemas/status'
        attempts:
          $ref: '#/components/schemas/attempts'
        parameters:
          $ref: '#/components/schemas/parameters'
      additionalProperties: false
    parameters:
      type: object
    attempts:
      type: number
    type:
      type: string
      maxLength: 250
    reason:
      type: string
    description:
      type: string
      maxLength: 2000
    creationTime:
      type: string
      format: date-time
    updateTime:
      type: string
      format: date-time
    jobId:
      type: string
      format: uuid
    taskId:
      type: string
      format: uuid
    percentage:
      type: integer
      minimum: 0
      maximum: 100
    status:
      type: string
      enum:
        - Pending
        - In-Progress
        - Completed
        - Failed
        - Expired
        - Aborted
      example: 'Pending'
...
    tasksResponse:
      type: array
      items:
        $ref: '#/components/schemas/taskResponse'
    taskResponse:
      type: object
      required:
        - id
        - jobId
        - description
        - created
        - updated
        - status
        - percentage
        - attempts
        - resettable
      properties:
        id:
          $ref: '#/components/schemas/taskId'
        jobId:
          $ref: '#/components/schemas/jobId'
        description:
          type: string
        parameters:
          type: object
        created:
          type: string
          format: date-time
        updated:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/status'
        percentage:
          $ref: '#/components/schemas/percentage'
        reason:
          type: string
        attempts:
          type: number
        resettable:
          type: boolean
...
    errorMessage:
      type: object
      properties:
        message::
          type: string
        stacktrace:
          type: string
      required:
        - message
    defaultOkMessage:
      type: object
      properties:
        code:
          type: string
          enum:
            - JOB_UPDATED_SUCCESSFULLY
            - JOB_DELETED_SUCCESSFULLY
            - JOB_RESET_SUCCESSFULLY
            - JOB_ABORTED_SUCCESSFULLY
            - TASK_UPDATED_SUCCESSFULLY
            - TASK_DELETED_SUCCESSFULLY
            - UPDATE_EXPIRED_STATUS_SUCCESSFULLY
          example: 'JOB_UPDATED_SUCCESSFULLY'
      required:
        - code
    createTaskPayload:
      type: object
      properties:
        description:
          type: string
          maxLength: 2000
        parameters:
          type: object
        reason:
          type: string
        percentage:
          $ref: '#/components/schemas/percentage'
        type:
          type: string
          maxLength: 255
        status:
          $ref: '#/components/schemas/status'
        attempts:
          type: integer
        blockDuplication:
          type: boolean
          default: false
      required:
        - parameters
        - type
      additionalProperties: false
    createTaskResponse:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/taskId'
    createMultipleTasksResponse:
      type: object
      properties:
        ids:
          type: array
          items:
            $ref: '#/components/schemas/taskId'
    updateTaskPayload:
      type: object
      properties:
        description:
          type: string
        parameters:
          type: object
        status:
          $ref: '#/components/schemas/status'
        percentage:
          $ref: '#/components/schemas/percentage'
        reason:
          type: string
        attempts:
          type: integer
      required:
        - status
      additionalProperties: false
    findInactiveTasksRequest:
      type: object
      properties:
        inactiveTimeSec:
          type: integer
        types:
          $ref: '#/components/schemas/taskTypes'
        ignoreTypes:
          $ref: '#/components/schemas/taskTypes'
      required:
        - inactiveTimeSec
    taskTypes:
      type: array
      items:
        $ref: '#/components/schemas/taskType'
    taskType:
      type: object
      properties:
        jobType:
          type: string
        taskType:
          type: string
    taskIdList:
      type: array
      items:
        $ref: '#/components/schemas/taskId'
...